requirements.txt
requests
html2text
pandas
transformers
gradio

numpy
google-generativeai
openai
